+++ crates/client/src/lib.rs
@@ -1,3 +1,4 @@
+#![cfg_attr(docsrs, feature(doc_cfg))]
 #![warn(missing_docs)] // error if there are missing docs
 
 //! This crate contains client implementations that can be used to contact the Temporal service.
@@ -20,6 +21,8 @@ pub mod grpc;
 pub mod interceptors;
 mod metrics;
 mod options_structs;
+#[cfg(feature = "experimental")]
+#[cfg_attr(docsrs, doc(cfg(feature = "experimental")))]
 /// Experimental APIs for configuring clients with reusable plugins.
 pub mod plugins;
 /// Visible only for tests
@@ -65,6 +68,8 @@ pub use interceptors::{
 };
 pub use metrics::{LONG_REQUEST_LATENCY_HISTOGRAM_NAME, REQUEST_LATENCY_HISTOGRAM_NAME};
 pub use options_structs::*;
+#[cfg(feature = "experimental")]
+#[cfg_attr(docsrs, doc(cfg(feature = "experimental")))]
 pub use plugins::{
     ClientPlugin, ErasedClientPlugin, PluginApplyError, PluginError, PluginTarget, WorkerPluginData,
 };
@@ -432,6 +437,14 @@ impl Connection {
         } else {
             None
         };
+        #[cfg(feature = "experimental")]
+        let payloads_warn_size = options.payload_limits.payloads_warn_size;
+        #[cfg(not(feature = "experimental"))]
+        let payloads_warn_size = 512 * 1024;
+        #[cfg(feature = "experimental")]
+        let memo_warn_size = options.payload_limits.memo_warn_size;
+        #[cfg(not(feature = "experimental"))]
+        let memo_warn_size = 2 * 1024;
         Ok(Self {
             inner: Arc::new(ConnectionInner {
                 service: svc_client,
@@ -445,12 +458,9 @@ impl Connection {
                 _dns_task: dns_task,
                 payloads_warn_size: resolve_warn_threshold(
                     "payloads_warn_size",
-                    options.payload_limits.payloads_warn_size,
+                    payloads_warn_size,
                 ),
-                memo_warn_size: resolve_warn_threshold(
-                    "memo_warn_size",
-                    options.payload_limits.memo_warn_size,
-                ),
+                memo_warn_size: resolve_warn_threshold("memo_warn_size", memo_warn_size),
             }),
         })
     }
@@ -1097,9 +1107,12 @@ impl Client {
     /// Connect to a Temporal service and create a namespace-bound client, applying registered
     /// plugins to connection and client options in registration order.
     pub async fn connect(
-        mut connection_options: ConnectionOptions,
+        connection_options: ConnectionOptions,
         client_options: ClientOptions,
     ) -> Result<Self, ClientConnectError> {
+        #[cfg(feature = "experimental")]
+        let mut connection_options = connection_options;
+        #[cfg(feature = "experimental")]
         plugins::apply_connection_plugins(&client_options, &mut connection_options)?;
         let connection = Connection::connect(connection_options).await?;
         Ok(Self::new(connection, client_options)?)
@@ -1109,7 +1122,10 @@ impl Client {
     ///
     /// Registered client plugins are applied here. Connection plugin hooks only run when using
     /// [`Client::connect`].
-    pub fn new(connection: Connection, mut options: ClientOptions) -> Result<Self, ClientNewError> {
+    pub fn new(connection: Connection, options: ClientOptions) -> Result<Self, ClientNewError> {
+        #[cfg(feature = "experimental")]
+        let mut options = options;
+        #[cfg(feature = "experimental")]
         plugins::apply_client_plugins(&mut options)?;
         Ok(Client {
             connection,
