+++ crates/client/src/plugins.rs
@@ -4,6 +4,7 @@ use crate::{ClientOptions, ConnectionOptions};
 use std::{any::Any, error::Error, sync::Arc};
 
 /// An error returned by a plugin configuration hook.
+#[cfg_attr(docsrs, doc(cfg(feature = "experimental")))]
 #[derive(Debug, thiserror::Error)]
 #[error(transparent)]
 pub struct PluginError(Box<dyn Error + Send + Sync>);
@@ -16,6 +17,7 @@ impl PluginError {
 }
 
 /// The configuration target being modified when a plugin failed.
+#[cfg_attr(docsrs, doc(cfg(feature = "experimental")))]
 #[derive(Clone, Copy, Debug, Eq, PartialEq, derive_more::Display)]
 #[non_exhaustive]
 pub enum PluginTarget {
@@ -34,6 +36,7 @@ pub enum PluginTarget {
 }
 
 /// An error applying a named plugin to a configuration target.
+#[cfg_attr(docsrs, doc(cfg(feature = "experimental")))]
 #[derive(Debug, thiserror::Error)]
 #[error("plugin '{plugin_name}' failed to configure {target}: {source}")]
 #[non_exhaustive]
@@ -64,6 +67,7 @@ impl PluginApplyError {
 /// Configures connection and namespace-bound client options.
 ///
 /// **Experimental:** This API may change or be removed.
+#[cfg_attr(docsrs, doc(cfg(feature = "experimental")))]
 pub trait ClientPlugin: Send + Sync + 'static {
     /// Return the stable, unique name used to identify this plugin in diagnostics and worker
     /// heartbeats.
@@ -87,6 +91,7 @@ pub trait ClientPlugin: Send + Sync + 'static {
 ///
 /// Implementing this trait will not make a type recognized as plugin. Only SDK known
 /// `WorkerPluginExtension` implementers will be used as plugins.
+#[cfg_attr(docsrs, doc(cfg(feature = "experimental")))]
 pub trait WorkerPluginData: Any + Send + Sync + 'static {}
 
 /// A type-erased client plugin and worker-plugin propagation data.
@@ -94,6 +99,7 @@ pub trait WorkerPluginData: Any + Send + Sync + 'static {}
 /// The worker data is intentionally opaque to avoid taking a dependency on `temporalio-sdk`.
 ///
 /// **Experimental:** This API may change or be removed.
+#[cfg_attr(docsrs, doc(cfg(feature = "experimental")))]
 #[derive(Clone)]
 pub struct ErasedClientPlugin {
     client: Arc<dyn ClientPlugin>,
@@ -174,7 +180,7 @@ pub(crate) fn apply_client_plugins(options: &mut ClientOptions) -> Result<(), Pl
     Ok(())
 }
 
-#[cfg(test)]
+#[cfg(all(test, feature = "experimental"))]
 mod tests {
     use super::*;
     use std::sync::atomic::{AtomicUsize, Ordering};
