+++ crates/sdk-core-c-bridge/Cargo.toml
@@ -42,6 +42,7 @@ xz2 = { version = "0.1", optional = true }
 [dependencies.temporalio-client]
 path = "../client"
 version = "~0.7.0"
+features = ["experimental"]
 
 [dependencies.temporalio-sdk-core]
 path = "../sdk-core"
