+++ crates/sdk/src/lib.rs
@@ -1,3 +1,4 @@
+#![cfg_attr(docsrs, feature(doc_cfg))]
 #![warn(missing_docs)] // error if there are missing docs
 
 //! This crate defines a Public Preview Temporal Rust SDK.
@@ -66,6 +67,8 @@ extern crate self as temporalio_sdk;
 pub mod activities;
 pub mod error;
 pub mod interceptors;
+#[cfg(feature = "experimental")]
+#[cfg_attr(docsrs, doc(cfg(feature = "experimental")))]
 /// Experimental APIs for configuring clients and workers with reusable plugins.
 pub mod plugins;
 pub mod runtime;
@@ -81,6 +84,11 @@ pub mod workflow_replayer;
 mod workflow_wasm;
 pub mod workflows;
 
+#[cfg(feature = "experimental")]
+#[cfg_attr(docsrs, doc(cfg(feature = "experimental")))]
+pub use crate::plugins::{
+    ClientAndWorkerPlugin, SimplePlugin, SimplePluginBuilder, SimplePluginOption, WorkerPlugin,
+};
 pub use crate::{
     error::{
         ActivityExecutionError, ApplicationFailure, ChildWorkflowExecutionError,
@@ -88,25 +96,25 @@ pub use crate::{
         RetryState, TimeoutType, WorkerCreateError, WorkerRunError, WorkerValidationError,
         WorkflowRegistrationError, WorkflowSignalError,
     },
-    plugins::{
-        ClientAndWorkerPlugin, SimplePlugin, SimplePluginBuilder, SimplePluginOption, WorkerPlugin,
-        WorkflowDefinitions,
-    },
+    workflow_registry::WorkflowDefinitions,
 };
 pub use runtime::Runtime;
 pub use temporalio_client::Namespace;
-#[cfg(feature = "experimental")]
-pub use temporalio_workflow::ContinueAsNewVersioningBehavior;
 pub use temporalio_workflow::{
     ActivityCancellationType, ActivityCloseTimeouts, ActivityOptions, BaseWorkflowContext,
     CancellableFuture, CancellableFutureWithReason, ChildWorkflowCancellationType,
     ChildWorkflowOptions, ContinueAsNewOptions, ExternalWorkflowHandle, LocalActivityOptions,
-    MemoValue, NexusOperationCancellationType, NexusOperationOptions, ParentClosePolicy,
-    PatchActivationCallback, SignalWorkflowOptions, StartChildWorkflowExecutionFailedCause,
-    StartChildWorkflowOutput, StartedChildWorkflow, StartedNexusOperation, SyncWorkflowContext,
-    TimerOptions, TimerResult, VersioningIntent, WaitConditionOptions, WorkflowCancellationError,
-    WorkflowCancellationToken, WorkflowContext, WorkflowContextView, WorkflowIdReusePolicy,
-    WorkflowRandomValue, WorkflowResult, WorkflowTermination,
+    MemoValue, ParentClosePolicy, SignalWorkflowOptions, StartChildWorkflowExecutionFailedCause,
+    StartChildWorkflowOutput, StartedChildWorkflow, SyncWorkflowContext, TimerOptions, TimerResult,
+    VersioningIntent, WaitConditionOptions, WorkflowCancellationError, WorkflowCancellationToken,
+    WorkflowContext, WorkflowContextView, WorkflowIdReusePolicy, WorkflowRandomValue,
+    WorkflowResult, WorkflowTermination,
+};
+#[cfg(feature = "experimental")]
+#[cfg_attr(docsrs, doc(cfg(feature = "experimental")))]
+pub use temporalio_workflow::{
+    ContinueAsNewVersioningBehavior, NexusOperationCancellationType, NexusOperationOptions,
+    PatchActivationCallback, PatchActivationInput, StartedNexusOperation,
 };
 #[cfg(feature = "wasm-workflows")]
 pub use workflow_wasm::WasmWorkflowComponent;
@@ -133,6 +141,8 @@ use std::{
     time::Duration,
 };
 use temporalio_client::{Client, ClientOptions, NamespacedClient};
+#[cfg(feature = "experimental")]
+use temporalio_common::protos::temporal::api::worker::v1::PluginInfo;
 use temporalio_common::{
     ActivityDefinition, WorkflowDefinition,
     data_converters::{
@@ -151,13 +161,14 @@ use temporalio_common::{
         },
         temporal::api::{
             common::v1::Payload, enums::v1::WorkflowTaskFailedCause, failure::v1::Failure,
-            worker::v1::PluginInfo,
         },
     },
     worker::{WorkerDeploymentOptions, WorkerTaskTypes, build_id_from_current_exe},
 };
 use temporalio_sdk_core::{PollError, init_worker};
-use temporalio_workflow::runtime::entry::WorkflowImplementation;
+use temporalio_workflow::{
+    InternalPatchActivationCallback, runtime::entry::WorkflowImplementation,
+};
 use tokio::sync::{
     Notify,
     mpsc::{UnboundedSender, unbounded_channel},
@@ -201,9 +212,11 @@ pub struct WorkerOptions {
     workflow_interceptor_constructors: Vec<WorkflowInterceptorConstructor>,
 
     #[builder(field)]
+    #[cfg(feature = "experimental")]
     worker_plugins: Vec<Arc<dyn WorkerPlugin>>,
 
     #[builder(field)]
+    #[cfg(feature = "experimental")]
     client_plugin_names: HashSet<String>,
 
     #[cfg(feature = "wasm-workflows")]
@@ -304,6 +317,15 @@ pub struct WorkerOptions {
     /// exceed the namespace error limits; oversized payloads are sent to server, which enforces the
     /// limit. Defaults to false.
     /// NOTE: Experimental
+    #[cfg(feature = "experimental")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "experimental")))]
+    #[cfg_attr(
+        docsrs,
+        builder(setters(
+            some_fn(name = __disable_payload_error_limit, vis = "pub(crate)"),
+            option_fn(name = __maybe_disable_payload_error_limit, vis = "pub(crate)")
+        ))
+    )]
     #[builder(default = false)]
     pub disable_payload_error_limit: bool,
     /// Experimental callback that decides whether the first non-replay call to
@@ -313,7 +335,67 @@ pub struct WorkerOptions {
     /// `true` records the patch marker; returning `false` leaves the patch inactive for the
     /// workflow run. For registered WASM workflow components, the callback remains on the worker
     /// host and is invoked through the workflow component's synchronous host interface.
+    #[cfg(feature = "experimental")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "experimental")))]
+    #[cfg_attr(
+        docsrs,
+        builder(setters(
+            some_fn(name = __patch_activation_callback, vis = "pub(crate)"),
+            option_fn(name = __maybe_patch_activation_callback, vis = "pub(crate)")
+        ))
+    )]
     pub patch_activation_callback: Option<PatchActivationCallback>,
+}
+
+#[cfg(all(feature = "experimental", docsrs))]
+impl<S: worker_options_builder::State> WorkerOptionsBuilder<S> {
+    /// Set whether payloads over the namespace error limit are sent to the server.
+    #[doc(cfg(feature = "experimental"))]
+    pub fn disable_payload_error_limit(
+        self,
+        value: bool,
+    ) -> WorkerOptionsBuilder<worker_options_builder::SetDisablePayloadErrorLimit<S>>
+    where
+        S::DisablePayloadErrorLimit: worker_options_builder::IsUnset,
+    {
+        self.__disable_payload_error_limit(value)
+    }
+
+    /// Set the payload error limit override from an optional value.
+    #[doc(cfg(feature = "experimental"))]
+    pub fn maybe_disable_payload_error_limit(
+        self,
+        value: Option<bool>,
+    ) -> WorkerOptionsBuilder<worker_options_builder::SetDisablePayloadErrorLimit<S>>
+    where
+        S::DisablePayloadErrorLimit: worker_options_builder::IsUnset,
+    {
+        self.__maybe_disable_payload_error_limit(value)
+    }
+
+    /// Set the callback used to decide whether a patch should activate.
+    #[doc(cfg(feature = "experimental"))]
+    pub fn patch_activation_callback(
+        self,
+        value: PatchActivationCallback,
+    ) -> WorkerOptionsBuilder<worker_options_builder::SetPatchActivationCallback<S>>
+    where
+        S::PatchActivationCallback: worker_options_builder::IsUnset,
+    {
+        self.__patch_activation_callback(value)
+    }
+
+    /// Set the patch activation callback from an optional value.
+    #[doc(cfg(feature = "experimental"))]
+    pub fn maybe_patch_activation_callback(
+        self,
+        value: Option<PatchActivationCallback>,
+    ) -> WorkerOptionsBuilder<worker_options_builder::SetPatchActivationCallback<S>>
+    where
+        S::PatchActivationCallback: worker_options_builder::IsUnset,
+    {
+        self.__maybe_patch_activation_callback(value)
+    }
 }
 
 impl<S: worker_options_builder::State> WorkerOptionsBuilder<S> {
@@ -338,6 +420,7 @@ impl<S: worker_options_builder::State> WorkerOptionsBuilder<S> {
         self
     }
 
+    #[cfg(feature = "experimental")]
     pub(crate) fn with_worker_plugins(
         mut self,
         worker_plugins: Vec<Arc<dyn WorkerPlugin>>,
@@ -358,12 +441,16 @@ impl<S: worker_options_builder::State> WorkerOptionsBuilder<S> {
     /// Register a worker plugin.
     ///
     /// **Experimental:** This API may change or be removed.
+    #[cfg(feature = "experimental")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "experimental")))]
     pub fn worker_plugin<P: WorkerPlugin>(mut self, plugin: P) -> Self {
         self.worker_plugins.push(Arc::new(plugin));
         self
     }
 
     /// Append a worker interceptor. Interceptors run in registration order.
+    #[cfg(feature = "experimental")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "experimental")))]
     pub fn worker_interceptor<I: WorkerInterceptor + 'static>(mut self, interceptor: I) -> Self {
         self.worker_interceptors.push(Arc::new(interceptor));
         self
@@ -476,6 +563,8 @@ fn def_build_id() -> WorkerDeploymentOptions {
 
 impl WorkerOptions {
     /// Append a worker interceptor. Interceptors run in registration order.
+    #[cfg(feature = "experimental")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "experimental")))]
     pub fn worker_interceptor<I: WorkerInterceptor + 'static>(
         &mut self,
         interceptor: I,
@@ -589,6 +678,25 @@ impl WorkerOptions {
         if !workflows_registered && !activities_registered {
             return Err("At least one workflow or activity must be registered".to_owned());
         }
+        #[cfg(feature = "experimental")]
+        let disable_payload_error_limit = self.disable_payload_error_limit;
+        #[cfg(not(feature = "experimental"))]
+        let disable_payload_error_limit = false;
+        #[cfg(feature = "experimental")]
+        let plugin_info = self
+            .client_plugin_names
+            .iter()
+            .map(|name| PluginInfo {
+                name: name.clone(),
+                version: String::new(),
+            })
+            .chain(self.worker_plugins.iter().map(|registration| PluginInfo {
+                name: registration.name().to_owned(),
+                version: String::new(),
+            }))
+            .collect();
+        #[cfg(not(feature = "experimental"))]
+        let plugin_info = HashSet::new();
 
         WorkerConfig::builder()
             .namespace(namespace)
@@ -627,20 +735,8 @@ impl WorkerOptions {
             ))
             .workflow_failure_errors(self.workflow_failure_errors.clone())
             .workflow_types_to_failure_errors(self.workflow_types_to_failure_errors.clone())
-            .plugins(
-                self.client_plugin_names
-                    .iter()
-                    .map(|name| PluginInfo {
-                        name: name.clone(),
-                        version: String::new(),
-                    })
-                    .chain(self.worker_plugins.iter().map(|registration| PluginInfo {
-                        name: registration.name().to_owned(),
-                        version: String::new(),
-                    }))
-                    .collect(),
-            )
-            .disable_payload_error_limit(self.disable_payload_error_limit)
+            .plugins(plugin_info)
+            .disable_payload_error_limit(disable_payload_error_limit)
             .build()
     }
 }
@@ -678,7 +774,7 @@ struct WorkflowHalf {
     workflow_removed_from_map: Notify,
     detect_nondeterministic_futures: bool,
     #[debug(skip)]
-    patch_activation_callback: Option<PatchActivationCallback>,
+    patch_activation_callback: Option<InternalPatchActivationCallback>,
 }
 #[derive(Debug)]
 struct WorkflowData {
@@ -784,8 +880,11 @@ impl Worker {
     pub fn new(
         runtime: &Runtime,
         client: Client,
-        mut options: WorkerOptions,
+        options: WorkerOptions,
     ) -> Result<Self, WorkerCreateError> {
+        #[cfg(feature = "experimental")]
+        let mut options = options;
+        #[cfg(feature = "experimental")]
         plugins::apply_worker_plugins(client.options(), &mut options)?;
         let wc = options
             .to_core_options(client.namespace(), client.identity())
@@ -817,8 +916,11 @@ impl Worker {
     pub fn new_from_core_options(
         worker: Arc<CoreWorker>,
         client_options: ClientOptions,
-        mut options: WorkerOptions,
+        options: WorkerOptions,
     ) -> Result<Self, WorkerCreateError> {
+        #[cfg(feature = "experimental")]
+        let mut options = options;
+        #[cfg(feature = "experimental")]
         plugins::apply_worker_plugins(&client_options, &mut options)?;
         Self::new_from_core_options_prepared(worker, client_options, options)
     }
@@ -847,7 +949,10 @@ impl Worker {
             workflow_interceptor_constructors,
         );
         me.set_detect_nondeterministic_futures(options.detect_nondeterministic_futures);
-        me.workflow_half.patch_activation_callback = options.patch_activation_callback;
+        #[cfg(feature = "experimental")]
+        {
+            me.workflow_half.patch_activation_callback = options.patch_activation_callback;
+        }
         #[cfg(feature = "wasm-workflows")]
         me.workflow_half
             .workflow_definitions
@@ -1749,30 +1854,6 @@ mod tests {
             .unwrap();
     }
 
-    #[test]
-    fn simple_plugin_workflow_function_merges_definitions() {
-        let plugin = SimplePlugin::builder("simple")
-            .workflows(|existing: Option<WorkflowDefinitions>| {
-                assert!(existing.is_some());
-                let mut workflows = WorkflowDefinitions::new();
-                workflows.register_workflow::<OtherWorkflow>().unwrap();
-                workflows
-            })
-            .build();
-        let client_options = ClientOptions::new("namespace").build();
-        let mut worker_options = WorkerOptions::new("task_q")
-            .register_workflow::<MyWorkflow>()
-            .unwrap()
-            .worker_plugin(plugin)
-            .build();
-
-        crate::plugins::apply_worker_plugins(&client_options, &mut worker_options).unwrap();
-
-        let workflows = format!("{:?}", worker_options.workflows());
-        assert!(workflows.contains("MyWorkflow"));
-        assert!(workflows.contains("OtherWorkflow"));
-    }
-
     #[rstest::rstest]
     #[case::workflow_only(true, false, Ok(WorkerTaskTypes::workflow_only()))]
     #[case::activity_only(false, true, Ok(WorkerTaskTypes::activity_only()))]
@@ -1908,24 +1989,6 @@ mod tests {
         assert_eq!(config.client_identity_override, expected);
     }
 
-    #[rstest::rstest]
-    #[case::default_enforces_error_limit(None, false)]
-    #[case::opt_out_disables_error_limit(Some(true), true)]
-    #[case::explicit_enable_error_limit(Some(false), false)]
-    #[test]
-    fn disable_payload_error_limit_propagates(
-        #[case] override_value: Option<bool>,
-        #[case] expected: bool,
-    ) {
-        let config = WorkerOptions::new("task_q")
-            .register_activities(MyActivities {})
-            .maybe_disable_payload_error_limit(override_value)
-            .build()
-            .to_core_options("ns".into(), String::new())
-            .unwrap();
-        assert_eq!(config.disable_payload_error_limit, expected);
-    }
-
     #[test]
     fn max_eager_activity_reservations_per_workflow_task_propagates() {
         let config = WorkerOptions::new("task_q")
@@ -1935,5 +1998,52 @@ mod tests {
             .to_core_options("ns".into(), String::new())
             .unwrap();
         assert_eq!(config.max_eager_activity_reservations_per_workflow_task, 7);
+    }
+
+    #[cfg(feature = "experimental")]
+    mod experimental_tests {
+        use super::*;
+
+        #[test]
+        fn simple_plugin_workflow_function_merges_definitions() {
+            let plugin = SimplePlugin::builder("simple")
+                .workflows(|existing: Option<WorkflowDefinitions>| {
+                    assert!(existing.is_some());
+                    let mut workflows = WorkflowDefinitions::new();
+                    workflows.register_workflow::<OtherWorkflow>().unwrap();
+                    workflows
+                })
+                .build();
+            let client_options = ClientOptions::new("namespace").build();
+            let mut worker_options = WorkerOptions::new("task_q")
+                .register_workflow::<MyWorkflow>()
+                .unwrap()
+                .worker_plugin(plugin)
+                .build();
+
+            crate::plugins::apply_worker_plugins(&client_options, &mut worker_options).unwrap();
+
+            let workflows = format!("{:?}", worker_options.workflows());
+            assert!(workflows.contains("MyWorkflow"));
+            assert!(workflows.contains("OtherWorkflow"));
+        }
+
+        #[rstest::rstest]
+        #[case::default_enforces_error_limit(None, false)]
+        #[case::opt_out_disables_error_limit(Some(true), true)]
+        #[case::explicit_enable_error_limit(Some(false), false)]
+        #[test]
+        fn disable_payload_error_limit_propagates(
+            #[case] override_value: Option<bool>,
+            #[case] expected: bool,
+        ) {
+            let config = WorkerOptions::new("task_q")
+                .register_activities(MyActivities {})
+                .maybe_disable_payload_error_limit(override_value)
+                .build()
+                .to_core_options("ns".into(), String::new())
+                .unwrap();
+            assert_eq!(config.disable_payload_error_limit, expected);
+        }
     }
 }
